fix: move isRedHatRegistryConfigured to src/utils.ts
