+++ src/rh-api/subscription.ts
@@ -17,39 +17,10 @@
  ***********************************************************************/
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 
-import { accessSync, constants, readFileSync } from 'node:fs';
-import { homedir } from 'node:os';
-import path from 'node:path';
-
 import createClient from 'openapi-fetch';
 
 import type { paths } from './gen/subscription';
 import { ClientHolder } from './utils';
-
-export const REGISTRY_REDHAT_IO = 'registry.redhat.io';
-
-// TODO: add listRegistries to registry API to allow search by
-// registry URL
-export function isRedHatRegistryConfigured(): boolean {
-  const pathToAuthJson = path.join(homedir(), '.config', 'containers', 'auth.json');
-  let configured = false;
-  try {
-    // TODO: handle all kind problems with file existence, accessibility and parsable content
-    accessSync(pathToAuthJson, constants.R_OK);
-    const authFileContent = readFileSync(pathToAuthJson, { encoding: 'utf8' });
-    const authFileJson: {
-      auths?: {
-        [registryUrl: string]: {
-          auth: string;
-        };
-      };
-    } = JSON.parse(authFileContent);
-    configured = authFileJson?.auths?.prototype.hasOwnProperty.call(authFileJson?.auths, REGISTRY_REDHAT_IO) || false;
-  } catch (_notAccessibleError) {
-    // if file is not there, ignore and return default value
-  }
-  return configured;
-}
 
 export class ActivationKey extends ClientHolder<paths> {
   async createActivationKeys(body: { name: string; role: string; usage: string; serviceLevel: string }) {
