+++ src/util.ts
@@ -16,7 +16,9 @@
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
 
-import { platform } from 'node:os';
+import { accessSync, constants, readFileSync } from 'node:fs';
+import { homedir, platform } from 'node:os';
+import path from 'node:path';
 
 import type { AuthenticationSession } from '@podman-desktop/api';
 import { authentication } from '@podman-desktop/api';
@@ -43,4 +45,27 @@ export async function signIntoRedHatDeveloperAccount(createIfNone = true): Promi
     ], // scope that gives access to console.redhat.com APIs
     { createIfNone },
   );
+}
+
+export const REGISTRY_REDHAT_IO = 'registry.redhat.io';
+
+export function isRedHatRegistryConfigured(): boolean {
+  const pathToAuthJson = path.join(homedir(), '.config', 'containers', 'auth.json');
+  let configured = false;
+  try {
+    // TODO: handle all kind problems with file existence, accessibility and parsable content
+    accessSync(pathToAuthJson, constants.R_OK);
+    const authFileContent = readFileSync(pathToAuthJson, { encoding: 'utf8' });
+    const authFileJson: {
+      auths?: {
+        [registryUrl: string]: {
+          auth: string;
+        };
+      };
+    } = JSON.parse(authFileContent);
+    configured = authFileJson?.auths?.prototype.hasOwnProperty.call(authFileJson?.auths, REGISTRY_REDHAT_IO) || false;
+  } catch (_notAccessibleError) {
+    // if file is not there, ignore and return default value
+  }
+  return configured;
 }
